bau: Move tests that update a gateway account
Specifically, move tests that call `PATCH`. Moving because
GatewayAccountResourceIT is becoming too big.
